+++ header/org/xresloader/Xresloader.java
@@ -11,9 +11,10 @@ public final class Xresloader {
     registry.add(org.xresloader.Xresloader.msgDescription);
     registry.add(org.xresloader.Xresloader.msgRequireMappingAll);
     registry.add(org.xresloader.Xresloader.msgSeparator);
-    registry.add(org.xresloader.Xresloader.verifier);
+    registry.add(org.xresloader.Xresloader.validator);
     registry.add(org.xresloader.Xresloader.fieldUniqueTag);
     registry.add(org.xresloader.Xresloader.fieldNotNull);
+    registry.add(org.xresloader.Xresloader.verifier);
     registry.add(org.xresloader.Xresloader.fieldDescription);
     registry.add(org.xresloader.Xresloader.fieldAlias);
     registry.add(org.xresloader.Xresloader.fieldRatio);
@@ -93,7 +94,7 @@ public final class Xresloader {
           .newFileScopedGeneratedExtension(
         java.lang.String.class,
         null);
-  public static final int VERIFIER_FIELD_NUMBER = 1001;
+  public static final int VALIDATOR_FIELD_NUMBER = 1001;
   /**
    * <pre>
    * 附加验证器
@@ -104,7 +105,7 @@ public final class Xresloader {
   public static final
     com.google.protobuf.GeneratedMessage.GeneratedExtension<
       com.google.protobuf.DescriptorProtos.FieldOptions,
-      java.lang.String> verifier = com.google.protobuf.GeneratedMessage
+      java.lang.String> validator = com.google.protobuf.GeneratedMessage
           .newFileScopedGeneratedExtension(
         java.lang.String.class,
         null);
@@ -137,6 +138,21 @@ public final class Xresloader {
       java.lang.Boolean> fieldNotNull = com.google.protobuf.GeneratedMessage
           .newFileScopedGeneratedExtension(
         java.lang.Boolean.class,
+        null);
+  public static final int VERIFIER_FIELD_NUMBER = 1009;
+  /**
+   * <pre>
+   * 附加验证器(兼容老名称)
+   * </pre>
+   *
+   * <code>extend .google.protobuf.FieldOptions { ... }</code>
+   */
+  public static final
+    com.google.protobuf.GeneratedMessage.GeneratedExtension<
+      com.google.protobuf.DescriptorProtos.FieldOptions,
+      java.lang.String> verifier = com.google.protobuf.GeneratedMessage
+          .newFileScopedGeneratedExtension(
+        java.lang.String.class,
         null);
   public static final int FIELD_DESCRIPTION_FIELD_NUMBER = 1011;
   /**
@@ -335,29 +351,30 @@ public final class Xresloader {
       "uf.MessageOptions\030\351\007 \001(\t:A\n\027msg_require_" +
       "mapping_all\022\037.google.protobuf.MessageOpt" +
       "ions\030\363\007 \001(\010:7\n\rmsg_separator\022\037.google.pr" +
-      "otobuf.MessageOptions\030\364\007 \001(\t:0\n\010verifier" +
-      "\022\035.google.protobuf.FieldOptions\030\351\007 \001(\t:8" +
-      "\n\020field_unique_tag\022\035.google.protobuf.Fie" +
-      "ldOptions\030\352\007 \003(\t:6\n\016field_not_null\022\035.goo" +
-      "gle.protobuf.FieldOptions\030\353\007 \001(\010:9\n\021fiel" +
-      "d_description\022\035.google.protobuf.FieldOpt" +
-      "ions\030\363\007 \001(\t:3\n\013field_alias\022\035.google.prot" +
-      "obuf.FieldOptions\030\364\007 \003(\t:3\n\013field_ratio\022" +
-      "\035.google.protobuf.FieldOptions\030\365\007 \001(\005:7\n" +
-      "\017field_separator\022\035.google.protobuf.Field" +
-      "Options\030\366\007 \001(\t:6\n\016field_required\022\035.googl" +
-      "e.protobuf.FieldOptions\030\367\007 \001(\010::\n\022field_" +
-      "origin_value\022\035.google.protobuf.FieldOpti" +
-      "ons\030\370\007 \001(\t:9\n\021oneof_description\022\035.google" +
-      ".protobuf.OneofOptions\030\351\007 \001(\t:7\n\017oneof_s" +
-      "eparator\022\035.google.protobuf.OneofOptions\030" +
-      "\352\007 \001(\t:6\n\016oneof_not_null\022\035.google.protob" +
-      "uf.OneofOptions\030\353\007 \001(\010:7\n\020enum_descripti" +
-      "on\022\034.google.protobuf.EnumOptions\030\351\007 \001(\t:" +
-      "=\n\021enumv_description\022!.google.protobuf.E" +
-      "numValueOptions\030\351\007 \001(\t:6\n\nenum_alias\022!.g" +
-      "oogle.protobuf.EnumValueOptions\030\352\007 \003(\tb\006" +
-      "proto3"
+      "otobuf.MessageOptions\030\364\007 \001(\t:1\n\tvalidato" +
+      "r\022\035.google.protobuf.FieldOptions\030\351\007 \001(\t:" +
+      "8\n\020field_unique_tag\022\035.google.protobuf.Fi" +
+      "eldOptions\030\352\007 \003(\t:6\n\016field_not_null\022\035.go" +
+      "ogle.protobuf.FieldOptions\030\353\007 \001(\010:0\n\010ver" +
+      "ifier\022\035.google.protobuf.FieldOptions\030\361\007 " +
+      "\001(\t:9\n\021field_description\022\035.google.protob" +
+      "uf.FieldOptions\030\363\007 \001(\t:3\n\013field_alias\022\035." +
+      "google.protobuf.FieldOptions\030\364\007 \003(\t:3\n\013f" +
+      "ield_ratio\022\035.google.protobuf.FieldOption" +
+      "s\030\365\007 \001(\005:7\n\017field_separator\022\035.google.pro" +
+      "tobuf.FieldOptions\030\366\007 \001(\t:6\n\016field_requi" +
+      "red\022\035.google.protobuf.FieldOptions\030\367\007 \001(" +
+      "\010::\n\022field_origin_value\022\035.google.protobu" +
+      "f.FieldOptions\030\370\007 \001(\t:9\n\021oneof_descripti" +
+      "on\022\035.google.protobuf.OneofOptions\030\351\007 \001(\t" +
+      ":7\n\017oneof_separator\022\035.google.protobuf.On" +
+      "eofOptions\030\352\007 \001(\t:6\n\016oneof_not_null\022\035.go" +
+      "ogle.protobuf.OneofOptions\030\353\007 \001(\010:7\n\020enu" +
+      "m_description\022\034.google.protobuf.EnumOpti" +
+      "ons\030\351\007 \001(\t:=\n\021enumv_description\022!.google" +
+      ".protobuf.EnumValueOptions\030\351\007 \001(\t:6\n\nenu" +
+      "m_alias\022!.google.protobuf.EnumValueOptio" +
+      "ns\030\352\007 \003(\tb\006proto3"
     };
     descriptor = com.google.protobuf.Descriptors.FileDescriptor
       .internalBuildGeneratedFileFrom(descriptorData,
@@ -368,21 +385,22 @@ public final class Xresloader {
     msgDescription.internalInit(descriptor.getExtensions().get(1));
     msgRequireMappingAll.internalInit(descriptor.getExtensions().get(2));
     msgSeparator.internalInit(descriptor.getExtensions().get(3));
-    verifier.internalInit(descriptor.getExtensions().get(4));
+    validator.internalInit(descriptor.getExtensions().get(4));
     fieldUniqueTag.internalInit(descriptor.getExtensions().get(5));
     fieldNotNull.internalInit(descriptor.getExtensions().get(6));
-    fieldDescription.internalInit(descriptor.getExtensions().get(7));
-    fieldAlias.internalInit(descriptor.getExtensions().get(8));
-    fieldRatio.internalInit(descriptor.getExtensions().get(9));
-    fieldSeparator.internalInit(descriptor.getExtensions().get(10));
-    fieldRequired.internalInit(descriptor.getExtensions().get(11));
-    fieldOriginValue.internalInit(descriptor.getExtensions().get(12));
-    oneofDescription.internalInit(descriptor.getExtensions().get(13));
-    oneofSeparator.internalInit(descriptor.getExtensions().get(14));
-    oneofNotNull.internalInit(descriptor.getExtensions().get(15));
-    enumDescription.internalInit(descriptor.getExtensions().get(16));
-    enumvDescription.internalInit(descriptor.getExtensions().get(17));
-    enumAlias.internalInit(descriptor.getExtensions().get(18));
+    verifier.internalInit(descriptor.getExtensions().get(7));
+    fieldDescription.internalInit(descriptor.getExtensions().get(8));
+    fieldAlias.internalInit(descriptor.getExtensions().get(9));
+    fieldRatio.internalInit(descriptor.getExtensions().get(10));
+    fieldSeparator.internalInit(descriptor.getExtensions().get(11));
+    fieldRequired.internalInit(descriptor.getExtensions().get(12));
+    fieldOriginValue.internalInit(descriptor.getExtensions().get(13));
+    oneofDescription.internalInit(descriptor.getExtensions().get(14));
+    oneofSeparator.internalInit(descriptor.getExtensions().get(15));
+    oneofNotNull.internalInit(descriptor.getExtensions().get(16));
+    enumDescription.internalInit(descriptor.getExtensions().get(17));
+    enumvDescription.internalInit(descriptor.getExtensions().get(18));
+    enumAlias.internalInit(descriptor.getExtensions().get(19));
     com.google.protobuf.DescriptorProtos.getDescriptor();
   }