+++ sample/proto_v2/kind_option.lua
@@ -119,7 +119,7 @@ local const_res = {
               name = "id",
               number = 2,
               options = {
-                verifier = "0-99999",
+                validator = "0-99999",
               },
               type_name = "UINT32",
             },
@@ -377,7 +377,7 @@ local const_res = {
               number = 3,
               options = {
                 field_description = "Refer to cost_type",
-                verifier = "cost_type",
+                validator = "cost_type",
               },
               type_name = "UINT32",
             },