+++ sample/proto_v3/kind_option.lua
@@ -114,7 +114,7 @@ local const_res = {
               name = "id",
               number = 2,
               options = {
-                verifier = "0-99999",
+                validator = "0-99999",
               },
               type_name = "UINT32",
             },
@@ -372,7 +372,7 @@ local const_res = {
               number = 3,
               options = {
                 field_description = "Refer to cost_type",
-                verifier = "cost_type",
+                validator = "cost_type",
               },
               type_name = "UINT32",
             },