+++ sample/proto_v3/kind_option.mod.lua
@@ -115,7 +115,7 @@ local const_res = {
               name = "id",
               number = 2,
               options = {
-                verifier = "0-99999",
+                validator = "0-99999",
               },
               type_name = "UINT32",
             },
@@ -373,7 +373,7 @@ local const_res = {
               number = 3,
               options = {
                 field_description = "Refer to cost_type",
-                verifier = "cost_type",
+                validator = "cost_type",
               },
               type_name = "UINT32",
             },